requirements.txt
Flask==2.3.3
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
scikit-learn==1.3.0
yfinance==0.2.28
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7